+++ execution_engine_testing/tests/src/test/regression/mod.rs
@@ -38,3 +38,4 @@ mod gh_2280;
 mod regression_20210707;
 mod regression_20210831;
 mod regression_20210924;
+mod transforms_must_be_ordered;